Cambiando diseño de menú y lista de pacientes a css

+++ src/app/components/citas/lista-citas/lista-citas.component.ts
@@ -1,12 +1,14 @@
 import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { Router, RouterLink } from '@angular/router';
+
 import { CitaService } from '../../../services/cita.service';
 import { Cita } from '../../../models/cita';
 
 @Component({
   selector: 'app-lista-citas',
   standalone: true,
-  imports: [CommonModule],
+  imports: [CommonModule, RouterLink],
   templateUrl: './lista-citas.component.html',
   styleUrl: './lista-citas.component.css'
 })
@@ -14,10 +16,18 @@ export class ListaCitasComponent implements OnInit {
 
   citas: Cita[] = [];
 
-  constructor(private citaService: CitaService) {}
+  constructor(
+    private citaService: CitaService,
+    private router: Router
+  ) {}
 
   ngOnInit(): void {
     this.cargarCitas();
+  }
+
+  editarCita(c: Cita) {
+  
+    this.router.navigate(['/citas/editar']);
   }
 
   cargarCitas() {
@@ -27,8 +37,10 @@ export class ListaCitasComponent implements OnInit {
   }
 
   eliminarCita(id: number) {
-    this.citaService.eliminarCita(id).subscribe(() => {
-      this.cargarCitas();
-    });
+    if (confirm('¿Desea cancelar esta cita médica de la clínica?')) {
+      this.citaService.eliminarCita(id).subscribe(() => {
+        this.cargarCitas();
+      });
+    }
   }
 }
